feat(SpanRow): Implement detailed span row components and utilities
- Add SpanRowDetailsCard for displaying detailed information about spans.
- Create SpanRowMainRow for rendering the main row of spans with interactive elements.
- Introduce SpanRow component to manage the rendering of main and detail rows.
- Implement utility functions for span search matching and action state retrieval.
- Add styled components for consistent UI representation of spans and their details.
- Create useSpanRowTheme hook to manage dark/light theme states.
